requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
python-dotenv>=1.0.1
celery>=5.3.6
redis>=5.0.1
requests>=2.31.0
google-generativeai>=0.3.3
